veri: use cranelift-codegen all-arch feature (#100)
This PR enables the `all-arch` feature for our dependency on the
`cranelift-codegen` crate.

I suspect this was the cause of confusing errors related to unused imports in
the codegen crate. By default `cranelift-codegen` only includes the host
architecture. Therefore, when switching from x86 to aarch64 (Linux workstation
to Apple laptop) it was possible that the build would work on one but not the
other.