isaspec: enable AluRRR mul/div opcodes (#32)
Enable the `*MulH` and `*Div` opcodes for the ASLp-generated `AluRRR`
specs.

Required implementing conversion for the ASLp `ite` builtin.

Unfortunately this inflates verification time quite a bit, I suspect
because the div opcodes introduce 128-bit ops. This incurs a general
slowdown, but also makes some of the harder cases hard enough that we
need to increase the solver timeout to get them to terminate. We'll
likely need to do something about this, perhaps using some version of
partial-specs that only outputs the active opcode branches.