isaspec: use substitution instead of equality constriants (#67)
When generating specs from ASLp, we have to define a mapping between
spec inputs and variables in the derived constraints. Previously we
would do this lazily by introducing equality constraints. This not only
looks awful #62, but anecdotally may also have solver performance
implications. This PR updates the spec generation to substitute the
mapped expressions for the variables instead.

Updates #62 #37