isaspec: generate multiple files (#69)
This PR changes the `isaspec` tool to generate multiple files.

As we extend the use of ASLp, we are generating a lot of spec code. This
is especially true for `BitRR` which we'll be adding soon. Allowing the
tool to generate more than one file will hopefully make them a bit more
manageable.

Updates #35 #62