spec: term tagging and filtering in verification (#72)
This PR adds support for tagging terms with syntax:

```
(attr <term> (tag <tag>))
```

In addition, the `veri` tool now allows for skipping expansions if they
contain a term with a given tag. Tagging may also be helpful in coverage
measurement #47.

Note that the explicit chaining syntax is updated to match:

```
(attr <term> (veri chain))
```

The `vector` tag is applied to vector instructions. This is intended to
help work to verify the `icmp` rules, which at the moment produces
expansions for vector types.

Updates #47