isle: unify provide/require spec parsing (#89)
This PR makes a minor tweak to the ISLE `spec` parsing to unify how
`provide` and `require` fields are parsed.

The intention is to make it slightly cleaner to parse additional fields,
such as to support state modeling #49. Doing this in a separate PR just
to verify this change is valid.