verifier: named type resolution (#97)
Implement logic to resolve all named references in a compound type.

The immediate motivation for this is to enable the use of named types in
term instantiations. At the moment named types will make it all the way
through to type inference, which will then fail on them. With this
change, types can be resolved prior to that point.