aarch64: model `Reg` as 64-bit (#108)
Change model of `Reg` on AArch64 to 64-bit rather than bit-vector of
unknown width. This is a reasonable assumption while we work on
lowerings to general purpose instructions, but will cause issues when we
want to verify vector registers.

The model of `Reg` is moved from the shared prelude to the
architecture-specific backends. The model of `Reg`.

Confirmed with a local run of `veri.sh` that the current AArch64 type
inference and verification results are unchanged.

Update #46