aarch64: verify limited subset of 64-bit load expansions (#111)
Introduces specification for CLIF load instructions, and verifies a
limited subset of the `I64` load lowerings.

Verifying these lowerings is a clear example of rule chaining at work.
We tag the `amode*` intermediate terms with the `chain` tag. As a result
the total number of expansions increases substantially to 9000+.

There are a 200+ expansions of the `load_i64` rule. In order to make
this work it was necessary to exclude a lot of terms for the time being
with a `TODO` tag, as well as model some types as unspecified. We will
need to return to these and specify them, or decide they are out of
scope.

Updates #49