verify: expand clif term instantiations (#125)
Some of the CLIF term specs had incomplete type instantiations. This PR
populates more complete lists.

For unclear reasons this triggers some of the verification solver calls
to take a long time on 16-bit instantiations. This would need more
investigation. For now, this PR drops the solver timeout to 10 seconds
so the timeouts don't take as long.

Updates #34