solver: unspecified model values (#146)
Add support for fetching models containing unspecified values.

With the addition of declared sorts in #141, this broke the code that
parses constants from SMT models. The backend solvers differ on how they
implement this. This PR deals with the CVC5 format only, in which
unspecified values are represented as something like `(as @Unspecified_0
Unspecified)`.

Updates #45