spec: matches clause (#150)
This PR adds a `matches` clause to specs which describes the conditions
when an extractor (non-infallible) or partial constructor matches.

Currently, the meaning of the `requires` clause on specs is conflated.
It's treated as a traditional function pre-condition, and as a
specification of when an `Option<..>` return from a term is `Some`. This
dual purpose leads to confusing semantics that prevents us from modeling
priorities correctly, see
https://github.com/mmcloughlin/veriisle-wasmtime/issues/128#issuecomment-2409113818.

This change may well be controversial given its departure from the
Crocus conventions. We could also revert to dual-purpose `require`
clauses, and change the semantics to fix the priority problem. I
personally would advocate for the new syntax as it clarifies the
meanings of the two cases.

Updates #128