verifier: support priorities (#151)
Provide support for incorporating rule priorities into verification
conditions.

When priorities are considered, the behavior is that for all strictly
higher-priority rules that may overlap with the rule under expansion,
the negation of the overlapping rule's constraints are brought into the
expansion. This necessitates brining in more bindings referenced by the
negated rule's constraints.

One problem is that overlap detection does not work when internal
extractors are left unexpanded. This is the most gross aspect of the
approach here, since we have to parse and run `sema` on the input ISLE
twice, once without internal extractor expansion just so we can deduce
the rule overlap sets.

When rule priorities are factored in it can lead to poor user experience
since the negated constraints bring in more terms that have to be
satisfied. In the fairly common case where priorities are not actually
relevant for correctness this can be annoying. Therefore, we introduce a
new rule attribute `(attr <rule> (veri priority))` which denotes that
the rule's correctness depends on priority. Negated higher-priority
rules will only be considered when this attribute is present.

Updates #128