aarch64: model `Reg` as unknown width (#172)
Model the `Reg` type on AArch64 as unknown bitvector width.

Reverses #108. The motivation for this is the need to model vector
intructions, where `Reg` is also used to represent ARM `Z` 128-bit
registers. The type inference problems that necessitated the 64-bit type
before have now been fixed by a combination of #167 and #171.

Updates #46 #34